pyproject.toml
[tool.poetry]
name = "chat-service"
version = "0.1.0"
description = ""
authors = ["TripHunter <info@triphunter.vn>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^1.0.1"
pydantic = "^2.7.4"
fastapi = {extras = ["all"], version = "^0.112.1"}
sqlalchemy = "^2.0.31"
alembic = "^1.13.1"
psycopg2 = "^2.9.9"
httpx = "^0.27.0"
typer = "^0.12.3"
sqlalchemy-utils = "^0.41.2"
pytest = "^8.2.2"
python-dateutil = "^2.9.0.post0"
langchain = "^0.2.5"
langchain-postgres = "^0.0.9"
pusher = "^3.3.2"
langchain-community = "^0.2.6"
langchain-openai = "^0.1.13"
pandas = "^2.2.2"
black = "23.3.0"

[black.tool]
line-length = 88
target-version = ['py310']
include = '\.pyi?$'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
